refactor: relocate refactor_primitive from engine to refactor (#8425)
engine/refactor_primitive.rs was not engine code — every function constructs
`refactor::auto::Insertion` objects tagged with `code_audit::AuditFinding`. It
imported only `crate::refactor::auto::*` and `crate::code_audit::AuditFinding`
(zero engine-local deps) and had a single consumer, `refactor::plan::generate::
builders`, which already lives in refactor.

Move it to refactor/primitive_builders.rs and repoint the one consumer. Pure
rename, no content change.

This severs the LAST engine -> code_audit edge AND the last engine -> refactor
edge simultaneously. After this, `engine` has zero dependencies on the audit or
refactor feature layers — the foundation is clean, unblocking `homeboy-engine`
extraction and, above it, `homeboy-audit` / `homeboy-deploy` / `homeboy-release`.

Refs #8425
