+++ crates/homeboy-core/src/refactor/plan/generate/builders.rs
@@ -1,4 +1,4 @@
-pub(crate) use crate::engine::refactor_primitive::{
+pub(crate) use crate::refactor::primitive_builders::{
     doc_line_removal, function_removal, insertion, manual_blocked, manual_only, new_file,
     range_removal, tagged_doc_line_removal, tagged_doc_reference_update, tagged_import_add,
     tagged_insertion as insertion_with_primitive, tagged_line_replacement, tagged_range_removal,
